feat: improve countdown page by splitting it up into sections
